+++ mocks/index.ts
@@ -1,0 +1,15 @@
+async function initMocks() {
+  if (typeof window === "undefined") {
+    const { server } = await import("./server");
+    server.listen({
+      onUnhandledRequest: "bypass",
+    });
+  } else {
+    const { worker } = await import("./browser");
+    worker.start();
+  }
+}
+
+initMocks();
+
+export {};