+++ next.config.js
@@ -1,8 +1,23 @@
 const path = require("path");
 
 module.exports = {
-  webpack: (config) => {
+  webpack: (config, { isServer }) => {
     config.resolve.alias["~"] = path.resolve(__dirname);
+
+    if (isServer) {
+      config.resolve.alias = {
+        ...config.resolve.alias,
+        "msw/browser": false,
+      };
+    } else {
+      // Setting `resolve.alias` to `false` will tell webpack to ignore a module.
+      // `msw/node` is a server-only module that exports methods not available in
+      // the `browser`.
+      config.resolve.alias = {
+        ...config.resolve.alias,
+        "msw/node": false,
+      };
+    }
 
     return config;
   },