+++ playwright.config.ts
@@ -1,4 +1,8 @@
 import { devices, PlaywrightTestConfig } from "@playwright/test";
+import dotenv from "dotenv";
+
+// uses the .env.test env file
+dotenv.config({ path: "./.env.testing" });
 
 const config: PlaywrightTestConfig = {
   testDir: "playwright/integration",
@@ -20,11 +24,13 @@ const config: PlaywrightTestConfig = {
   reporter: "html",
 
   use: {
+    // headless: true,
     // Base URL to use in actions like `await page.goto('/')`.
     baseURL: "http://localhost:3000",
 
-    // Collect trace when retrying the failed test.
-    trace: "on-first-retry",
+    /* When running tests locally, record a trace for each test, but remove it from successful runs.
+     * On CI, turn this feature off. See https://playwright.dev/docs/trace-viewer */
+    trace: process.env.CI ? "off" : "retain-on-failure",
   },
   // Configure projects for major browsers.
   projects: [