+++ playwright/support/routes.ts
@@ -1,3 +1,0 @@
-export const NYPL_LOGIN_URL = "https://login.nypl.org/auth/login?redirect_uri=";
-export const LIMITED_ACCESS_WORK_PATH =
-  "/work/97496f55-2017-4fc3-9143-c593d36c8c30?featured=6969309";