+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - SFR-2076: Fix DRB PW Regression Tests
 - Update README with instructions to run app locally
 - Update `butlerlogic/action-autotag` version to be pinned since the stable version is using an unsupported Node version
+- Add Playwright integration test for cookie validation
 
 ## [0.18.1]